# glotcast-api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/me": {
      "get": {
        "operationId": "MeController_get_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      },
      "patch": {
        "operationId": "MeController_update_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      },
      "delete": {
        "operationId": "MeController_remove_v1",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "me"
        ]
      }
    },
...
    "/v1/admin/me": {
      "get": {
        "operationId": "AdminDashboardController_me_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMeDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/dashboard": {
      "get": {
        "operationId": "AdminDashboardController_get_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DashboardDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/categories": {
      "get": {
        "operationId": "AdminCategoriesController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCategoriesDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "post": {
        "operationId": "AdminCategoriesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCategoryDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/categories/{id}": {
      "patch": {
        "operationId": "AdminCategoriesController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminCategoryDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminCategoriesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/audit": {
      "get": {
        "operationId": "AdminAuditController_list_v1",
        "parameters": [
          {
            "name": "action",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 60,
              "type": "string"
            }
          },
          {
            "name": "targetId",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 80,
              "type": "string"
            }
          },
          {
            "name": "actorId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "integer"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 9007199254740991,
              "exclusiveMinimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminAuditDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    }
  },
  "info": {
    "title": "GlotCast API",
    "description": "Podcasts in three English levels: catalog, listening, vocabulary, translation, and the admin content pipeline. The contract is docs/contract-v1.md; `admin`-tagged operations are for the admin panel only.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "EpisodeSummary_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "podcast": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "name": {
                "type": "string"
              },
              "coverUrl": {
                "anyOf": [
                  {
                    "type": "string",
                    "minLength": 0
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "id",
              "name",
              "coverUrl"
            ],
            "additionalProperties": false
          },
          "number": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "type": "string"
          },
          "coverUrl": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "bannerUrl": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "isPro": {
            "type": "boolean"
          },
          "publishedAt": {
            "type": "string"
          },
          "levels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EpisodeLevelSummary_Output"
            },
            "description": "sorted bg, in, ad; only levels that have audio"
          }
        },
        "required": [
          "id",
          "podcast",
          "number",
          "title",
          "coverUrl",
          "bannerUrl",
          "isPro",
          "publishedAt",
          "levels"
        ],
        "additionalProperties": false
      },
      "EpisodeLevelSummary_Output": {
        "type": "object",
        "properties": {
          "level": {
            "$ref": "#/components/schemas/Level_Output"
          },
          "durationSec": {
            "type": "number"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "level",
          "durationSec",
          "description"
        ],
        "additionalProperties": false
      },
      "Level_Output": {
        "type": "string",
        "enum": [
          "bg",
          "in",
          "ad"
        ],
        "description": "bg = beginner, in = intermediate, ad = advanced"
      },
...
      "MeDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "isAnonymous": {
            "type": "boolean"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "avatarUrl": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "nativeLanguage": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ],
            "description": "BCP-47-ish code, e.g. \"tr\", \"pt-BR\""
          },
          "uiLanguage": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "translationLanguage": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "level": {
            "$ref": "#/components/schemas/Level_Output"
          },
          "dailyGoalMin": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "interests": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "motivation": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "career",
                  "travel",
                  "exams",
                  "fun",
                  "other"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "reminderTime": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ],
            "description": "\"HH:mm\", local time"
          },
          "featureAccess": {
            "type": "boolean",
            "description": "backend-granted Pro"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "isAnonymous",
          "email",
          "name",
          "avatarUrl",
          "nativeLanguage",
          "uiLanguage",
          "translationLanguage",
          "level",
          "dailyGoalMin",
          "interests",
          "motivation",
          "reminderTime",
          "featureAccess",
          "createdAt"
        ],
        "additionalProperties": false
      },
      "Level": {
        "type": "string",
        "enum": [
          "bg",
          "in",
          "ad"
        ],
        "description": "bg = beginner, in = intermediate, ad = advanced"
      },
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ]
          },
          "nativeLanguage": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z]{2,3}(-[A-Za-z0-9]{2,8})*$"
              },
              {
                "type": "null"
              }
            ]
          },
          "uiLanguage": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z]{2,3}(-[A-Za-z0-9]{2,8})*$"
              },
              {
                "type": "null"
              }
            ]
          },
          "translationLanguage": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z]{2,3}(-[A-Za-z0-9]{2,8})*$"
              },
              {
                "type": "null"
              }
            ]
          },
          "level": {
            "$ref": "#/components/schemas/Level"
          },
          "dailyGoalMin": {
            "type": "integer",
            "minimum": 1,
            "maximum": 600
          },
          "interests": {
            "maxItems": 30,
            "type": "array",
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 40
            }
          },
          "motivation": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "career",
                  "travel",
                  "exams",
                  "fun",
                  "other"
                ]
              },
              {
                "type": "null"
              }
            ]
          },
          "reminderTime": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "AdminMeDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "email"
        ],
        "additionalProperties": false
      },
      "DayCount_Output": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "date",
          "count"
        ],
        "additionalProperties": false
      },
      "DashboardDto_Output": {
        "type": "object",
        "properties": {
          "users": {
            "type": "object",
            "properties": {
              "total": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "anonymous": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "last7Days": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991,
                "description": "created in the last 7 days"
              }
            },
            "required": [
              "total",
              "anonymous",
              "last7Days"
            ],
            "additionalProperties": false
          },
          "dau": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DayCount_Output"
            },
            "description": "distinct listeners per (local) day, the last 30 days, oldest first"
          },
          "episodes": {
            "type": "object",
            "properties": {
              "published": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "drafts": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991,
                "description": "drafts and scheduled"
              }
            },
            "required": [
              "published",
              "drafts"
            ],
            "additionalProperties": false
          },
          "topEpisodes": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "episode": {
                  "$ref": "#/components/schemas/EpisodeSummary_Output"
                },
                "listeners": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                }
              },
              "required": [
                "episode",
                "listeners"
              ],
              "additionalProperties": false
            },
            "description": "most distinct listeners over the last 30 days"
          }
        },
        "required": [
          "users",
          "dau",
          "episodes",
          "topEpisodes"
        ],
        "additionalProperties": false
      },
...
      "AdminCategory_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "coverUrl": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "position": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "display order (ascending)"
          },
          "podcastCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "all podcasts, drafts included"
          },
          "createdAt": {
            "type": "string",
            "description": "ISO timestamp"
          },
          "updatedAt": {
            "type": "string",
            "description": "ISO timestamp"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "coverUrl",
          "position",
          "podcastCount",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "AdminCategoriesDto_Output": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/AdminCategory_Output"
        }
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "slug": {
            "description": "default: from the name",
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ]
          },
          "coverUrl": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ]
          },
          "position": {
            "type": "integer",
            "minimum": 0,
            "maximum": 10000
          }
        },
        "required": [
          "name"
        ]
      },
      "AdminCategoryDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "coverUrl": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "position": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "display order (ascending)"
          },
          "podcastCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "all podcasts, drafts included"
          },
          "createdAt": {
            "type": "string",
            "description": "ISO timestamp"
          },
          "updatedAt": {
            "type": "string",
            "description": "ISO timestamp"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "coverUrl",
          "position",
          "podcastCount",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "slug": {
            "description": "default: from the name",
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ]
          },
          "coverUrl": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ]
          },
          "position": {
            "type": "integer",
            "minimum": 0,
            "maximum": 10000
          }
        }
      },
...
      "AdminAuditDto_Output": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "at": {
                  "type": "string"
                },
                "actor": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid",
                          "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                        },
                        "email": {
                          "anyOf": [
                            {
                              "type": "string",
                              "minLength": 0
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "email"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "description": "the admin; null = the CLI (role changes)"
                },
                "action": {
                  "type": "string",
                  "description": "episode.create · episode.publish · level.put · user.update · role.grant …"
                },
                "targetType": {
                  "anyOf": [
                    {
                      "type": "string",
                      "minLength": 0
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "targetId": {
                  "anyOf": [
                    {
                      "type": "string",
                      "minLength": 0
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "meta": {
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {}
                }
              },
              "required": [
                "id",
                "at",
                "actor",
                "action",
                "targetType",
                "targetId",
                "meta"
              ],
              "additionalProperties": false
            }
          },
          "nextBefore": {
            "type": "integer",
            "description": "pass as ?before= for older entries; absent at the end",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "entries"
        ],
        "additionalProperties": false
      }
    }
  }
}